+++ docs/court_wizard.d.ts
@@ -6,15 +6,15 @@ export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembl
 export interface InitOutput {
   readonly memory: WebAssembly.Memory;
   readonly main: (a: number, b: number) => number;
-  readonly __wasm_bindgen_func_elem_204176: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_203975: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_197351: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_197362: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_204174: (a: number, b: number, c: number, d: number) => void;
-  readonly __wasm_bindgen_func_elem_199477: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_199497: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_204188: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_204178: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_204243: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_204040: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_204241: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_204239: (a: number, b: number, c: number, d: number) => void;
+  readonly __wasm_bindgen_func_elem_204253: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_197416: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_197427: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_199542: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_199562: (a: number, b: number) => void;
   readonly __wbindgen_export: (a: number, b: number) => number;
   readonly __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
   readonly __wbindgen_export3: (a: number) => void;