+++ docs/court_wizard_bg.wasm.d.ts
@@ -2,15 +2,15 @@
 /* eslint-disable */
 export const memory: WebAssembly.Memory;
 export const main: (a: number, b: number) => number;
-export const __wasm_bindgen_func_elem_204176: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_203975: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_197351: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_197362: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_204174: (a: number, b: number, c: number, d: number) => void;
-export const __wasm_bindgen_func_elem_199477: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_199497: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_204188: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_204178: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_204243: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_204040: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_204241: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_204239: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_204253: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_197416: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_197427: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_199542: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_199562: (a: number, b: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;