+++ src/game/loading/spawn_queue.rs
@@ -26,6 +26,7 @@ pub enum SpawnTask {
     #[allow(dead_code)]
     Castle,
     Wizard,
+    LoadCauldronAssets,
     Cauldron,
     PathfindingGrid,
 }